Move test_control to test lib (#989)
* test_control in util
